fix: adjustments handleValues
