+++ prisma/schema.prisma
@@ -26,15 +26,15 @@ model User {
   lastName      String?
   email         String  @unique
   password      String
-  salary        Float   @default(0)
-  moneySaved    Float   @default(0)
-  totalSpent    Float   @default(0)
-  totalExpenses Float   @default(0)
-  sessionsId    String
-  flagsId       String
+  salary        Float?
+  moneySaved    Float?
+  totalSpent    Float?
+  totalExpenses Float?
+  sessionsId    String?
+  flagsId       String?
 
-  sessions Sessions   @relation(fields: [sessionsId], references: [id])
-  flags    Flags      @relation(fields: [flagsId], references: [id])
+  sessions Sessions?  @relation(fields: [sessionsId], references: [id])
+  flags    Flags?     @relation(fields: [flagsId], references: [id])
   spending Spending[]
   expenses Expenses[]
   earnings Earnings[]
