+++ src/types/IValues.ts
@@ -1,0 +1,15 @@
+export interface IHandleValues {
+  value?: number
+  userTotal?: number | null
+  bodyValue: number
+}
+
+export interface IHandleSubtractExpenses {
+  value?: number
+  userTotal?: number | null
+}
+
+export interface IHandleSumValues {
+  userValue?: number | null
+  bodyValue: number
+}
